Added Debug Logs

+++ pkg/schema/types.go
@@ -1,8 +1,8 @@
 package schema
 
 type EventResult struct {
-	ID int
-	EventId string
-	E2ELatency int
-	EventType string
+	ID         int
+	EventId    string
+	E2ELatency int64
+	EventType  string
 }
